Fixed multimodify missing cache after error message.
The Back button on the error message page will send the user back to the current page, but the information which item and what data the user entered was lost.
Implemented some cache handling to allow the user go back to correct attribute multimodify.

Also fixed the case, if the field contains the replace mode, so it will select the correct mode.