Fixed permission system if NConf is not running under webroot path. (as normal user)
The new permission system did not work if user installed NConf in subdirectorys like:
http://example.com/some/folders/nconf

This should now be fixed but needs some further testing.
Only "user" accounts were affected, because admin is superuser and not restricted by the permission system.
---
The permission system is new, can handle the navigation links itself and is going to be more improved in the future...