Implemented further checks for the Permission system.
Problem was: The URL check was not enough. The users could show, or delete items by just manipulating the IDs in the URL.

I fixed these with additional checks.

Now the permission system will check for requesting IDs and have a look if the user has access to the class which they belong to.
Now its not possible to have a look on an item, which the user has no access to.