FIXED: Default values of checkcommand not used in advanced services -> http://forum.nconf.org/viewtopic.php?f=16&t=771
Exactly it implements the default checkcommands when on "add service" or "add advanced-service" and changing the "check command".
This will load the "default command params" of the check command into the "service parameters (params for check command)

Simplified code a bit: created "get default checkcommand params"-Function. (which also is needed on ajax/service_add.php